requirements.txt
grpcio==1.38.0
grpcio-tools==1.38.0
protobuf==3.17.3
six==1.16.0
